+++ activerecord/test/cases/enum_test.rb
@@ -700,6 +700,20 @@ class EnumTest < ActiveRecord::TestCase
     assert_not_predicate computer, :"Etc/GMT-1?"
   end
 
+  test "deserialize enum value to original hash key" do
+    proposed = Struct.new(:to_s).new("proposed")
+    written = Struct.new(:to_s).new("written")
+    klass = Class.new(ActiveRecord::Base) do
+      self.table_name = "books"
+      enum status: { proposed => 0, written => 1 }
+    end
+
+    book = klass.create!(status: 0)
+    assert_equal proposed, book.status
+    assert_predicate book, :proposed?
+    assert_not_predicate book, :written?
+  end
+
   test "enum logs a warning if auto-generated negative scopes would clash with other enum names" do
     old_logger = ActiveRecord::Base.logger
     logger = ActiveSupport::LogSubscriber::TestHelper::MockLogger.new